Remove Namespace component; rename TreeComponent to ModelBrowser
We have only one component to be concerned about now.
