+++ gaphor/ui/tests/test_modelbrowser.py
@@ -1,27 +1,22 @@
 import pytest
-from gi.repository import Gtk
 
 from gaphor import UML
 from gaphor.core.modeling import Diagram
-from gaphor.ui.treecomponent import (
+from gaphor.ui.modelbrowser import (
     ElementDragData,
     SearchEngine,
-    TreeComponent,
+    ModelBrowser,
     get_first_selected_item,
     list_item_drop_drop,
 )
 
-skip_if_gtk3 = pytest.mark.skipif(
-    Gtk.get_major_version() == 3, reason="Gtk.ListView is not supported by GTK 3"
-)
-
 
 @pytest.fixture
-def tree_component(event_manager, element_factory, modeling_language):
-    tree_component = TreeComponent(event_manager, element_factory, modeling_language)
-    tree_component.open()
-    yield tree_component
-    tree_component.close()
+def model_browser(event_manager, element_factory, modeling_language):
+    model_browser = ModelBrowser(event_manager, element_factory, modeling_language)
+    model_browser.open()
+    yield model_browser
+    model_browser.close()
 
 
 class ItemChangedHandler:
@@ -36,24 +31,22 @@ class ItemChangedHandler:
         self.added += added
 
 
-@skip_if_gtk3
-def test_show_item_in_tree_list_model(tree_component, element_factory):
+def test_show_item_in_tree_list_model(model_browser, element_factory):
     class_ = element_factory.create(UML.Class)
     package = element_factory.create(UML.Package)
 
     class_.package = package
 
-    pos = tree_component.select_element(class_)
-    tree_model = tree_component.selection
+    pos = model_browser.select_element(class_)
+    tree_model = model_browser.selection
 
     assert pos == 1
     assert tree_model.get_item(0).get_item().element is package
     assert tree_model.get_item(1).get_item().element is class_
 
 
-@skip_if_gtk3
-def test_tree_component_remove_element(tree_component, element_factory):
-    tree_model = tree_component.model.root
+def test_model_browser_remove_element(model_browser, element_factory):
+    tree_model = model_browser.model.root
     element = element_factory.create(UML.Class)
     items_changed = ItemChangedHandler()
     tree_model.connect("items-changed", items_changed)
@@ -64,12 +57,11 @@ def test_tree_component_remove_element(tree_component, element_factory):
     assert items_changed.removed == 1
 
 
-@skip_if_gtk3
-def test_tree_subtree_changed(tree_component, element_factory):
+def test_tree_subtree_changed(model_browser, element_factory):
     class_ = element_factory.create(UML.Class)
     package = element_factory.create(UML.Package)
 
-    tree_model = tree_component.model
+    tree_model = model_browser.model
     root_model = tree_model.root
     root_model_changed = ItemChangedHandler()
     root_model.connect("items-changed", root_model_changed)
@@ -81,9 +73,8 @@ def test_tree_subtree_changed(tree_component, element_factory):
     assert root_model_changed.removed == 3  # remove + node changed
 
 
-@skip_if_gtk3
-def test_tree_component_add_nested_element(tree_component, element_factory):
-    tree_model = tree_component.model
+def test_model_browser_add_nested_element(model_browser, element_factory):
+    tree_model = model_browser.model
     class_ = element_factory.create(UML.Class)
     package = element_factory.create(UML.Package)
 
@@ -95,9 +86,8 @@ def test_tree_component_add_nested_element(tree_component, element_factory):
     assert tree_model.tree_item_for_element(class_) in child_model
 
 
-@skip_if_gtk3
-def test_tree_component_unset_nested_element(tree_component, element_factory):
-    tree_model = tree_component.model
+def test_model_browser_unset_nested_element(model_browser, element_factory):
+    tree_model = model_browser.model
     class_ = element_factory.create(UML.Class)
     package = element_factory.create(UML.Package)
 
@@ -112,9 +102,8 @@ def test_tree_component_unset_nested_element(tree_component, element_factory):
     assert child_model is None
 
 
-@skip_if_gtk3
-def test_tree_component_remove_nested_element(tree_component, element_factory):
-    tree_model = tree_component.model
+def test_model_browser_remove_nested_element(model_browser, element_factory):
+    tree_model = model_browser.model
     class_ = element_factory.create(UML.Class)
     package = element_factory.create(UML.Package)
 
@@ -129,10 +118,9 @@ def test_tree_component_remove_nested_element(tree_component, element_factory):
     assert child_model is None
 
 
-@skip_if_gtk3
-def test_element_name_changed(tree_component, element_factory):
+def test_element_name_changed(model_browser, element_factory):
     class_ = element_factory.create(UML.Class)
-    tree_item = tree_component.model.tree_item_for_element(class_)
+    tree_item = model_browser.model.tree_item_for_element(class_)
 
     class_.name = "foo"
 
@@ -143,20 +131,18 @@ def test_element_name_changed(tree_component, element_factory):
     assert style.as_int().value == 0
 
 
-@skip_if_gtk3
-def test_element_weight_changed(tree_component, element_factory):
+def test_element_weight_changed(model_browser, element_factory):
     diagram = element_factory.create(Diagram)
-    tree_item = tree_component.model.tree_item_for_element(diagram)
+    tree_item = model_browser.model.tree_item_for_element(diagram)
     weight, style = tree_item.attributes.get_attributes()
 
     assert weight.as_int().value == 700
     assert style.as_int().value == 0
 
 
-@skip_if_gtk3
-def test_element_style_changed(tree_component, element_factory):
+def test_element_style_changed(model_browser, element_factory):
     class_ = element_factory.create(UML.Class)
-    tree_item = tree_component.model.tree_item_for_element(class_)
+    tree_item = model_browser.model.tree_item_for_element(class_)
 
     class_.isAbstract = True
     weight, style = tree_item.attributes.get_attributes()
@@ -165,50 +151,46 @@ def test_element_style_changed(tree_component, element_factory):
     assert style.as_int().value == 2
 
 
-@skip_if_gtk3
-def test_tree_component_model_ready(event_manager, element_factory, modeling_language):
+def test_model_browser_model_ready(event_manager, element_factory, modeling_language):
     class_ = element_factory.create(UML.Class)
     package = element_factory.create(UML.Package)
     class_.package = package
-    tree_component = TreeComponent(event_manager, element_factory, modeling_language)
-    tree_component.open()
-    tree_model = tree_component.model
+    model_browser = ModelBrowser(event_manager, element_factory, modeling_language)
+    model_browser.open()
+    tree_model = model_browser.model
 
     element_factory.model_ready()
 
     assert tree_model.tree_item_for_element(package) is not None
     assert tree_model.tree_item_for_element(class_) is None
 
-    tree_component.close()
+    model_browser.close()
 
 
-@skip_if_gtk3
-def test_tree_model_expand_to_relationship(tree_component, element_factory):
+def test_tree_model_expand_to_relationship(model_browser, element_factory):
     association = element_factory.create(UML.Association)
     package = element_factory.create(UML.Package)
     association.package = package
 
-    pos = tree_component.select_element(association)
+    pos = model_browser.select_element(association)
 
     assert pos == 2
-    assert tree_component.get_selected_element() is association
+    assert model_browser.get_selected_element() is association
 
 
-@skip_if_gtk3
-def test_create_diagram(tree_component, element_factory):
-    tree_component.tree_view_create_diagram("my type")
+def test_create_diagram(model_browser, element_factory):
+    model_browser.tree_view_create_diagram("my type")
 
     diagram = next(element_factory.select(Diagram))
 
     assert diagram.diagramType == "my type"
 
 
-@skip_if_gtk3
-def test_create_package(tree_component, element_factory):
+def test_create_package(model_browser, element_factory):
     parent = element_factory.create(UML.Package)
     parent.name = "root"
-    tree_component.select_element(parent)
-    tree_component.tree_view_create_package()
+    model_browser.select_element(parent)
+    model_browser.tree_view_create_package()
 
     package = next(p for p in element_factory.select(UML.Package) if p.name != "root")
 
@@ -216,9 +198,8 @@ def test_create_package(tree_component, element_factory):
     assert package.owner is parent
 
 
-@skip_if_gtk3
-def test_create_toplevel_package(tree_component, element_factory):
-    tree_component.tree_view_create_package()
+def test_create_toplevel_package(model_browser, element_factory):
+    model_browser.tree_view_create_package()
 
     package = next(p for p in element_factory.select(UML.Package) if p.name != "root")
 
@@ -226,51 +207,47 @@ def test_create_toplevel_package(tree_component, element_factory):
     assert package.owner is None
 
 
-@skip_if_gtk3
-def test_delete_element(tree_component, element_factory):
+def test_delete_element(model_browser, element_factory):
     klass = element_factory.create(UML.Class)
-    tree_component.select_element(klass)
-    assert tree_component.get_selected_element() is klass
+    model_browser.select_element(klass)
+    assert model_browser.get_selected_element() is klass
 
-    tree_component.tree_view_delete()
+    model_browser.tree_view_delete()
 
     assert not element_factory.lselect()
 
 
-@skip_if_gtk3
-def test_search_next(tree_component, element_factory):
+def test_search_next(model_browser, element_factory):
     class_a = element_factory.create(UML.Class)
     class_a.name = "a"
     class_b = element_factory.create(UML.Class)
     class_b.name = "b"
 
-    search_engine = SearchEngine(tree_component.model, tree_component.tree_view)
-    tree_component.select_element(class_a)
-    assert tree_component.get_selected_element() is class_a
+    search_engine = SearchEngine(model_browser.model, model_browser.tree_view)
+    model_browser.select_element(class_a)
+    assert model_browser.get_selected_element() is class_a
 
     search_engine.search_next("b")
 
-    assert tree_component.get_selected_element() is class_b
+    assert model_browser.get_selected_element() is class_b
 
 
-@skip_if_gtk3
-def test_search_text_changed(tree_component, element_factory):
+def test_search_text_changed(model_browser, element_factory):
     class_a = element_factory.create(UML.Class)
     class_a.name = "a"
     class_b = element_factory.create(UML.Class)
     class_b.name = "b"
 
-    search_engine = SearchEngine(tree_component.model, tree_component.tree_view)
-    tree_component.select_element(class_a)
-    assert tree_component.get_selected_element() is class_a
+    search_engine = SearchEngine(model_browser.model, model_browser.tree_view)
+    model_browser.select_element(class_a)
+    assert model_browser.get_selected_element() is class_a
 
     search_engine.text_changed("b")
 
-    assert tree_component.get_selected_element() is class_b
+    assert model_browser.get_selected_element() is class_b
 
 
-@skip_if_gtk3
-def test_generalization_text(tree_component, element_factory):
+def test_generalization_text(model_browser, element_factory):
     general = element_factory.create(UML.Class)
     general.name = "General"
     specific = element_factory.create(UML.Class)
@@ -279,7 +256,7 @@ def test_generalization_text(tree_component, element_factory):
     generalization.specific = specific
     generalization.general = general
 
-    model = tree_component.model
+    model = model_browser.model
     tree_item = model.tree_item_for_element(specific)
     branch = model.branches[tree_item]
     assert tree_item
@@ -287,8 +264,7 @@ def test_generalization_text(tree_component, element_factory):
     assert branch.relationships[0].text == "general: General"
 
 
-@skip_if_gtk3
-def test_drop_multiple_elements(tree_component, element_factory, event_manager):
+def test_drop_multiple_elements(model_browser, element_factory, event_manager):
     class_a = element_factory.create(UML.Class)
     class_a.name = "a"
     class_b = element_factory.create(UML.Class)
@@ -302,14 +278,14 @@ def test_drop_multiple_elements(tree_component, element_factory, event_manager):
     package = element_factory.create(UML.Package)
 
     drag_data = ElementDragData(elements=[class_a, class_b])
-    tree_component.select_element(class_a)
-    list_item = MockRowItem(get_first_selected_item(tree_component.selection))
-    tree_component.select_element(package)
+    model_browser.select_element(class_a)
+    list_item = MockRowItem(get_first_selected_item(model_browser.selection))
+    model_browser.select_element(package)
     target = MockDropTarget()
     list_item_drop_drop(target, drag_data, 0, 0, list_item, event_manager)
 
-    assert class_b not in tree_component.get_selected_elements()
-    assert class_a not in tree_component.get_selected_elements()
+    assert class_b not in model_browser.get_selected_elements()
+    assert class_a not in model_browser.get_selected_elements()
 
 
 class MockRowItem:
